requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
cachetools==5.3.2
slowapi==0.1.8
pydantic-settings==2.1.0 